refactor: 이미지 업로드 함수 분리
